huanhuan

+++ oa/src/com/oa/controller/system/PermissionController.java
@@ -21,6 +21,7 @@ import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import com.oa.bean.Permission;
 import com.oa.bean.ResponseResult;
+import com.oa.bean.RolePermission;
 import com.oa.bean.User;
 import com.oa.service.system.PermissionService;
 
@@ -199,18 +200,50 @@ public class PermissionController {
 	@ResponseBody
 	public ResponseResult deletePermission(@PathVariable("ids") String ids) {
 		ResponseResult rr =new ResponseResult();
+		List<RolePermission> rolePermissionList = null;
+		List<Permission> listPermission = null;
 		if(ids.contains("-")) {
 			String[] split_ids = ids.split("-");
 			List<Integer> listId = new ArrayList<Integer>();
 			for(String id : split_ids) {
 				listId.add(Integer.parseInt(id));
+				rolePermissionList = permissionService.getRolePermissionList(Integer.parseInt(id));
+				if(rolePermissionList!=null) {
+					rr.setStateCode(0);
+					rr.setMessage("菜单已在使用中。。。");
+					return rr;
+				}
+				
+				listPermission = permissionService.getParPermission(Integer.parseInt(id));
+				if(listPermission!=null) {
+					rr.setStateCode(0);
+					rr.setMessage("该菜单已有子菜单，您无法删除");
+					return rr;
+				}
+				
 			}
 			permissionService.deletePermissionBatch(listId);
+			return rr.success();
 		}else {
 			Integer id = Integer.parseInt(ids);
+			rolePermissionList = permissionService.getRolePermissionList(id);
+			if(rolePermissionList!=null) {
+				rr.setStateCode(0);
+				rr.setMessage("菜单已在使用中。。。");
+				return rr;
+			}
+			
+			listPermission = permissionService.getParPermission(id);
+			if(listPermission!=null) {
+				rr.setStateCode(0);
+				rr.setMessage("该菜单已有子菜单，您无法删除");
+				return rr;
+			}
+			listPermission = permissionService.getParPermission(id);
 			permissionService.deletePermission(id);
+			return rr.success();
 		}
-		return rr.success();
+		
 	}
 
 }
